14/08

+++ src/main/java/vn/eledevo/vksbe/service/customer/CustomerServiceImpl.java
@@ -8,14 +8,21 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import vn.eledevo.vksbe.dto.request.CustomerRequest;
+import vn.eledevo.vksbe.dto.response.ApiResponse;
 import vn.eledevo.vksbe.dto.response.CustomerResponse;
 import vn.eledevo.vksbe.entity.Customer;
+import vn.eledevo.vksbe.entity.Employee;
 import vn.eledevo.vksbe.entity.Order;
+import vn.eledevo.vksbe.exception.ValidationException;
 import vn.eledevo.vksbe.mapper.CustomerMapper;
 import vn.eledevo.vksbe.repository.CustomerRepository;
+import vn.eledevo.vksbe.repository.EmployeeRepository;
 import vn.eledevo.vksbe.repository.OrderRepository;
 
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static vn.eledevo.vksbe.constant.ResponseMessage.*;
 
 @Service
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
@@ -24,19 +31,24 @@ public class CustomerServiceImpl implements CustomerService{
 
     OrderRepository orderRepository;
 
+    EmployeeRepository employeeRepository;
+
     CustomerRepository customerRepository;
 
     CustomerMapper customerMapper;
 
     @Override
-    public List<CustomerResponse> getAllCustomer() {
+    public ApiResponse<List<CustomerResponse>> getAllCustomer() {
         List<Customer> customerList = customerRepository.findAll();
-        List<CustomerResponse> customerResponseList = customerMapper.toListResponse(customerList);
-        return customerResponseList;
+        List<CustomerResponse> customerResponseList = customerList
+                .stream()
+                .map(customerMapper::mapEntityToResponse)
+                .collect(Collectors.toList());
+        return new ApiResponse<>(200, "Lấy dữ liệu thành công", customerResponseList);
     }
 
     @Override
-    public List<CustomerResponse> getCustomer(Customer textSearch, String sortField, String sortDirection, int currentPage, int limitPage) {
+    public ApiResponse<List<CustomerResponse>> getCustomer(Customer textSearch, String sortField, String sortDirection, int currentPage, int limitPage) {
         Sort.Direction direction = Sort.Direction.ASC;
         if (sortDirection.equalsIgnoreCase("desc")) {
             direction = Sort.Direction.DESC;
@@ -50,36 +62,69 @@ public class CustomerServiceImpl implements CustomerService{
                 textSearch.getStatus(),
                 pageable
         );
-        List<CustomerResponse> customerResponseList = customerMapper.toListResponse(customerList);
-        return customerResponseList;
+        List<CustomerResponse> customerResponseList = customerList
+                .stream()
+                .map(customerMapper::mapEntityToResponse)
+                .collect(Collectors.toList());
+        return new ApiResponse<>(200, "Lấy dữ liệu thành công", customerResponseList);
     }
 
     @Override
-    public Customer addCustomer(CustomerRequest customerRequest) {
-        Customer customer = customerMapper.toEntity(customerRequest);
-        return customerRepository.save(customer);
+    public ApiResponse<Customer> addCustomer(CustomerRequest customerRequest) throws ValidationException{
+        Customer customer = customerMapper.mapRequestToEntity(customerRequest);
+        if (customerRequest.getEmployee() == null){
+            customer.setEmployee(null);
+        } else if (customerRepository.existsByEmployeeId(customerRequest.getEmployee())){
+            throw new ValidationException("Employee", EMPLOYEE_EXIST);
+        } else if (!employeeRepository.existsById(customerRequest.getEmployee())){
+            throw new ValidationException("Employee", EMPLOYEE_NOT_EXIST);
+        } else {
+            Employee employee = employeeRepository.findById(customerRequest.getEmployee()).get();
+            customer.setEmployee(employee);
+        }
+        return new ApiResponse<>(200, "Thêm dữ liệu thành công", customerRepository.save(customer));
     }
 
     @Override
-    public Customer updateCustomer(Long id, CustomerRequest customerRequest) {
+    public ApiResponse<Customer> updateCustomer(Long id, CustomerRequest customerRequest) throws ValidationException {
+        if (!customerRepository.existsById(id)){
+            throw new ValidationException("Customer", CUSTOMER_NOT_EXIST);
+        }
         Customer customer = customerRepository.findById(id).get();
+        if (customerRequest.getEmployee() == null){
+            customer.setEmployee(null);
+        } else if (customerRepository.existsByEmployeeId(customerRequest.getEmployee())){
+            throw new ValidationException("Employee", EMPLOYEE_EXIST);
+        } else if (!employeeRepository.existsById(customerRequest.getEmployee())){
+            throw new ValidationException("Employee", EMPLOYEE_NOT_EXIST);
+        } else {
+            Employee employee = employeeRepository.findById(customerRequest.getEmployee()).get();
+            customer.setEmployee(employee);
+        }
         customer.setName(customerRequest.getName());
         customer.setAddress(customerRequest.getAddress());
         customer.setPhone(customerRequest.getPhone());
         customer.setStatus(customerRequest.getStatus());
-        customer.setEmployee(customerRequest.getEmployee());
-        return customerRepository.save(customer);
+        List<Order> orders = orderRepository.findByCustomerId(id);
+        for (Order order : orders) {
+            order.setCustomerName(customerRequest.getName());
+            orderRepository.save(order);
+        }
+        return new ApiResponse<>(200, "Cập nhật dữ liệu thành công", customerRepository.save(customer));
     }
 
     @Override
-    public Customer deleteCustomer(Long id) {
+    public ApiResponse<Customer> deleteCustomer(Long id) throws ValidationException {
+        if (!customerRepository.existsById(id)){
+            throw new ValidationException("Customer", CUSTOMER_NOT_EXIST);
+        }
         List<Order> orders = orderRepository.findByCustomerId(id);
         for (Order order : orders) {
             order.setCustomer(null);
             orderRepository.save(order);
         }
         customerRepository.deleteById(id);
-        return null;
+        return new ApiResponse<>(200, "Xóa dữ liệu thành công");
     }
 
 }
